modify report type management

+++ frontend/client/src/components/ReportTypesManagement/index.js
@@ -16,57 +16,67 @@ import {
     Input,
     FormLabel,
     useToast,
-    Heading
+    Heading,
+    Icon
 } from '@chakra-ui/react'
+import { FaTrashAlt } from "react-icons/fa";
+import { IoSearchOutline } from "react-icons/io5";
+import { CiCirclePlus } from "react-icons/ci";
 
 import { useEffect } from "react";
 import axios from "axios";
 import Pagination from "../ReportDashboard/Pagination"
 import ReportProcessForm from "../ReportProcessForm";
+import { FaEye,FaPen } from "react-icons/fa";
+import { TbReportSearch } from "react-icons/tb";
+import { useNavigate, useLocation } from 'react-router-dom';
+
+import ToolkitProvider, { Search } from 'react-bootstrap-table2-toolkit/dist/react-bootstrap-table2-toolkit';
+import BootstrapTable from 'react-bootstrap-table-next';
+import paginationFactory from 'react-bootstrap-table2-paginator';
 
 import { BsFillTrashFill, BsFillPencilFill } from "react-icons/bs";
 const ITEMS_PER_PAGE = 10;
 const ReportTypesManagement = () => {
     //console.log(userData);
     //const { userData } = useUser();
+    const navigate = useNavigate();
     const [reportTypes, setReportTypes] = useState([]);
-
-    useEffect(() => {
-        const fetchReport = async () => {
-            try {
-                const response = await axios.get('http://127.0.0.1:5000/api/v1/reportTypes');
-                setReportTypes(response.data.data);
-            } catch (error) {
-                console.error('Error fetching report:', error.message);
-            }
-        };
-
-        fetchReport();
-    }, []);
-
-
-
-    //console.log(report);
     const { isOpen: isEditModalOpen, onOpen: onEditModalOpen, onClose: onEditModalClose } = useDisclosure();
     const { isOpen: isDelModalOpen, onOpen: onDelModalOpen, onClose: onDelModalClose } = useDisclosure();
     const { isOpen: isAddModalOpen, onOpen: onAddModalOpen, onClose: onAddModalClose } = useDisclosure();
 
     const [selectedReportType, setSelectedReportType] = useState(null);
     const [name, setName] = useState('')
+    const [update, setUpdate] = useState(true);
     const toast = useToast();
     const handleEdit = async (reportType) => {
         try {
-            const response = await axios.put(`http://127.0.0.1:5000/api/v1/reportTypes/${reportType._id}`, { name: name });
+            var isExist = null;
+            console.log(reportType);
+            const apiCheck = await axios.post(`http://127.0.0.1:5000/api/v1/reportTypes/findRpType`, { area: name });
+            if(apiCheck.data.data.length === 0){
+                const response = await axios.put(`http://127.0.0.1:5000/api/v1/reportTypes/${reportType._id}`, { name: name });
+                const apiUpdate = await axios.post(`http://127.0.0.1:5000/api/v1/report/updateRpType`,{oldRpType: reportType.name, newRpType: name})
+                setUpdate(true);
+                isExist = false
+            }
+            else isExist = true;
 
+            if(isExist === false) {
+                setName('')
+                setTimeout(() => {
+                    onEditModalClose();
+                }, 1000);
+            }        
             toast({
-                title: "Update successful.",
-                description: "You've successfully updated.",
-                status: "success",
+                title:(isExist ? ('Error') : ('Successful.')),
+                description:(isExist? ("Loại báo cáo đã tồn tại") : ("Loại báo cáo đã được cập nhật.")),
+                status:(isExist ? ('error') : ('success')),
                 duration: 2000,
                 isClosable: true,
             });
-            onEditModalClose();
-            window.location.reload();
+            
         } catch (error) {
             console.log(error);
         }
@@ -75,17 +85,30 @@ const ReportTypesManagement = () => {
 
     const handleDelete = async (reportType) => {
         try {
-            const response = await axios.delete(`http://127.0.0.1:5000/api/v1/reportTypes/${reportType._id}`);
+            var isContain = null;
+            const apiContain = await axios.post(`http://127.0.0.1:5000/api/v1/report/findByRpType`, {area: `${reportType.name}`});
+
+            if(apiContain.data.data.length === 0){
+                const response = await axios.delete(`http://127.0.0.1:5000/api/v1/reportTypes/${reportType._id}`);
+                setUpdate(true);
+                isContain = false;
+            }
+            else isContain = true;
+
+            if(isContain === false) {
+                setTimeout(() => {
+                    onDelModalClose();
+                }, 1000);
+            }
 
             toast({
-                title: "Delete successful.",
-                description: "You've successfully deleted.",
-                status: "success",
+                title: isContain? ('Error') : ('Successful.'),
+                description:isContain ? ("Loại báo cáo đã tồn tại báo cáo") : ("Loại báo cáo đã được xóa."),
+                status:isContain ? ('error') : ('success'),
                 duration: 2000,
                 isClosable: true,
-            });
-            onDelModalClose();
-            window.location.reload();
+            }); 
+            
         } catch (error) {
             console.log(error);
         }
@@ -93,30 +116,197 @@ const ReportTypesManagement = () => {
 
     const handleAddNew = async () => {
         try {
-            const response = await axios.post(`http://127.0.0.1:5000/api/v1/reportTypes/`, { name: name });
-
+            var isExist = null;
+            const apiCheck = await axios.post(`http://127.0.0.1:5000/api/v1/reportTypes/findRpType`, { area: name });
+            if(apiCheck.data.data.length === 0){
+                const response = await axios.post(`http://127.0.0.1:5000/api/v1/reportTypes/`, { name: name });
+                setUpdate(true);
+                isExist = false
+            }
+            else isExist = true
+            if(isExist === false) {
+                setName('')
+                setTimeout(() => {
+                    onAddModalClose();
+                }, 1000);
+            }        
             toast({
-                title: "Add successful.",
-                description: "You've successfully added.",
-                status: "success",
+                title:(isExist ? ('Error') : ('Successful.')),
+                description:(isExist? ("Loại báo cáo đã tồn tại") : ("Loại báo cáo đã được thêm mới.")),
+                status:(isExist ? ('error') : ('success')),
                 duration: 2000,
                 isClosable: true,
             });
-            onAddModalClose();
-            window.location.reload();
+
         } catch (error) {
             console.log(error);
         }
     }
 
 
+    useEffect(() => {
+        const fetchReport = async () => {
+            try {
+                const response = await axios.get('http://127.0.0.1:5000/api/v1/reportTypes');
+                setReportTypes(response.data.data);
+                setUpdate(false);
+            } catch (error) {
+                console.error('Error fetching report:', error.message);
+            }
+        };
+
+        fetchReport();
+    }, [update]);
+
+    const columns = [
+        {
+          dataField: 'name',
+          text: 'Report Type'
+        }, 
+        {
+          dataField: 'action',
+          isDummyField: true,
+          text: 'Action',
+          formatter: (cellContent, row) => {
+            return(
+              <div style={{display:"flex", alignItems:"center"}}>
+              {/* <Icon 
+                variant="unstyled" 
+                as={FaEye} 
+                w={5} 
+                h={5} 
+                marginRight={5} 
+                // marginLeft={2} 
+                onClick={
+                  () => navigate('/manage-ward',{state: { district: row?.name }})
+                }
+                _hover={{color:'blue'}}
+              /> */}
+              {/* <Icon 
+                as={TbReportSearch} 
+                w={5}
+                h={5}
+                marginRight={5}
+                onClick={
+                  () => navigate('/table-advertising-type',{state: { adType: row?.name }})
+                }
+                _hover={{color:'blue'}}
+              /> */}
+              <Icon 
+                as={FaPen} 
+                w={4} 
+                h={4}
+                marginRight={5}
+                onClick={
+                  () => {
+                    setSelectedReportType(row)
+                    onEditModalOpen()
+                  }
+                }
+                _hover={{color:'blue'}}
+              />
+              <Icon 
+                as={FaTrashAlt} 
+                w={4} 
+                h={4}
+                marginRight={5}
+                onClick={
+                  () => {
+                    setSelectedReportType(row)
+                    onDelModalOpen() 
+                  }
+                }
+                _hover={{color:'red'}}
+              />
+              </div>
+            )
+          }
+        }
+    ];
+
+    const sizePerPageOptionRenderer = ({
+        text,
+        page,
+        onSizePerPageChange
+    }) => (
+        <li
+          key={ text }
+          role="presentation"
+          className="dropdown-item"
+        >
+          <a
+            href="#"
+            tabIndex="-1"
+            role="menuitem"
+            data-page={ page }
+            onMouseDown={ (e) => {
+              e.preventDefault();
+              onSizePerPageChange(page);
+            } }
+            style={ { display:"block",color: 'red', width:"100%" } }
+          >
+            { text }
+          </a>
+        </li>
+    );
+
+    const options = {
+        sizePerPageOptionRenderer,
+        alwaysShowAllBtns: true,
+        hidePageListOnlyOnePage: true,
+        sizePerPageList: [{
+          text: '10', value: 10
+        }, {
+          text: '15', value: 15
+        }, {
+          text: '20', value: 20
+        }],
+    }
+
+
+    const MySearch = (props) => {
+        return (
+            <div className="form-group has-search">
+              <span className="form-control-feedback">
+                <Icon as={IoSearchOutline}></Icon>
+              </span>
+              <input
+              className="form-control"
+              style={{marginTop:"20px", marginBottom:"15px"}}
+              type="text"
+              onChange={(event) => {
+                props.onSearch(event.target.value)
+              }}
+              placeholder='Search'
+              />
+              <div style={{width:"100%",display:"flex",justifyContent:"flex-end"}}>
+              <Button leftIcon={<CiCirclePlus size="20" />} justifyContent="flex-start" width="200px" colorScheme='teal' variant='solid' 
+                onClick={() =>{
+                    onAddModalOpen()
+                  }
+                }>
+                Add
+              </Button>
+              </div>
+            </div>
+        );
+    };
+
+    const CaptionElement = () => <h3 style={{ borderRadius: '0.25em', textAlign: 'center', color: 'purple', border: '1px solid purple', padding: '0.5em', marginTop:"15px" }}>Report Type Management</h3>;
+
+    
+    
+
+    //console.log(report);
+    
+
     const [currentPage, setCurrentPage] = useState(1);
     const indexOfLastItem = currentPage * ITEMS_PER_PAGE;
     const indexOfFirstItem = indexOfLastItem - ITEMS_PER_PAGE;
     const currentItems = reportTypes.slice(indexOfFirstItem, indexOfLastItem);
     return (
-        <Box style={{ width: "100%", height: "100vh" }}>
-            <Heading mx={5}>
+        <div style={{width:"95%"}}>
+            {/* <Heading mx={5}>
                 Quản lý loại hình báo cáo
                 <Button ml={5}
 
@@ -129,8 +319,8 @@ const ReportTypesManagement = () => {
 
                                 Thêm mới
                             </Button>
-            </Heading>
-            <Table colorScheme='green'>
+            </Heading> */}
+            {/* <Table colorScheme='green'>
                 <Thead>
                     <Tr>
                         <Th>ID</Th>
@@ -178,7 +368,36 @@ const ReportTypesManagement = () => {
                 currentPage={currentPage}
                 totalPages={Math.ceil(reportTypes.length / ITEMS_PER_PAGE)}
                 onPageChange={setCurrentPage}
-            />
+            /> */}
+            <div style={{width:"100%"}}>
+                {reportTypes ? (
+                    <ToolkitProvider
+                        keyField="id"
+                        data={ reportTypes }
+                        columns={ columns }
+                        search
+                    >
+                        {
+                        props => (
+                            <div>
+                            <CaptionElement/>
+                            <MySearch 
+                                { ...props.searchProps } 
+                            />
+                            <BootstrapTable
+                                { ...props.baseProps }
+                                pagination={paginationFactory(options)} 
+                                striped
+                                rowStyle={{verticalAlign:"middle"}}
+                                bordered= {false}
+                            />
+                            </div>
+                        )
+                        }
+                    </ToolkitProvider>
+                    ) : (null)
+                }
+            </div>
             <Modal isOpen={isAddModalOpen} onClose={onAddModalClose} size='4xl'>
                         <ModalOverlay />
                         <ModalContent>
@@ -266,22 +485,22 @@ const ReportTypesManagement = () => {
                             <ModalBody>
 
                                 <FormLabel>Bạn chắc chắn muốn xóa?</FormLabel>
-                                <Button
-
-                                    variant='outline'
-                                    colorScheme='red'
-                                    onClick={() => {
-                                        handleDelete(selectedReportType)
-                                    }}
-                                >
-
-                                    Lưu
-                                </Button>
-
+                                
 
                             </ModalBody>
                             <ModalFooter>
-                                
+                            <Button
+
+                            variant='outline'
+                            colorScheme='red'
+                            onClick={() => {
+                                handleDelete(selectedReportType)
+                            }}
+                            >
+
+                            Yes
+                            </Button>
+
                             </ModalFooter>
 
                         </ModalContent>
@@ -289,7 +508,7 @@ const ReportTypesManagement = () => {
                 </Box>
 
             )}
-        </Box>
+        </div>
     );
 };
